ikt: First check-in
iKT is a set of command line tools, including a curses UI,
meant to help administrate and monitor Kubernetes clusters.
At present it's not a full-fledged solution--for several operations
it relies on kubeadm/kubectl, and it's definitely still a work
in progress.

Bug reports, feature requests, and suggestions for improvements
are warmly welcome.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>