iktadm: Switch context before configuring CNI
When configuring CNI we already have a running cluster.
Make sure that we're interacting with the right one--we might
have multiple clusters configured in .kube/config

Also, don't do sanity checking for the phase we're at before
we check for --list-tasks; we want to be able to list available
tasks for a command even if we're not able to run that command
at the moment (for instance we might want to check steps for
setup even if we haven't run prepare yet).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>