iku: Implement support for local playbooks
The distro playbooks can be overriden by having a playbook
with the same name in a local playbook directory. The easiest
way to modify the behaviour of a distro playbook is thus to
copy the distro playbook to your local playbook directory
and then modifying it.

If you want to completely disable a distro playbook you can
add an empty file with the same name as the distro playbook.
Empty files are silently ignored.

Beware that disabling internal playbooks (notably get_facts,
get_versions; possibly others) is likely to cause undefined
behaviour.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>