logparser: Support for drop-in parsers
Add support for drop-in parsers, and implement kubernetes-dashboard
and metrics-scraper as examples. Long term all parsers should be
split out.

To avoid performance issues the parsers shouldn't be read
every time though; it's definitely necessary to initialise
the parser list once on startup (and possibly allow for
regenerate the list, to avoid having to restart when new
parser definitions are added).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>