logparser,*: Major logparser changes
Some improvements for the YAML/JSON parser (affecting both
the ConfigMap viewer and the log viewer), added a parser file
for kube-apiserver (CAVEAT: not tested with older versions
of Kubernetes), and fixed a long-standing bug regarding to
severity for error messages in structured log messages.

Various behaviour of the key=value and YAML/JSON parsers
have also been moved to logparser_configuration; so far this
isn't actually exposed to the user, but it's a start at least.
Ideally these options should probably be possible to toggle
per parser rather than globally.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>