logparser: Parser file for starboard
Convert log parsing for starboard into a parser file.
While at it fix the parser to handle line-by-line YAML formatting.
To fully support the various formats supported by Starboard
a new flag had to be added to logparser() where the container
type is specified; this is because an init container can
have the same name as a container (or, as in this case,
neither of them have names that can be matched on requiring
the image to be the match, but the init container has different
log format than the container).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>