kubernetes_helper: Abort if server unreachable
The check for available APIs is, to say the least, very inefficient;
we iterate over all APIs we know, and check whether or not they
are available. This is obviously not ideal. Especially if
the server isn't responding; in such a case we wait for server
timeout for every single API we have defined. That's a LOT of timeouts.

With this change we now abort get_available_api_families() as soon
as get_api_resources() returns 503 (Service Unavailable)
or 42503 (our own artificial return value that signifies that
the server isn't responding).

This makes things fail a lot faster. The handling is still not ideal
though (even when disregarding the suboptimal API check); no error
message explaining the lack of resources is returned or displayed,
instead the user just faces a mostly empty list of views; some of
which are still going to be problematic).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>