iku: Add support for severity prefixes
Per default we strip severity prefixes from all messages
and visualise severity using colour instead. This is a good default,
but from an accessibility point of view (or when using a console
with a limited palette) it's not a good idea.

Add a configuration option to Add standardised severity prefixes.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>