iku: Initial support for deprecation information
Add a view that displays deprecated APIs that have been requested,
and what version they'll be deprecated in.

The next step is to have a corresponding info view that lists
what components are responsible for requesting them.

This will make it easier to find out what components, if any,
need to be updated before migrating to a new Kubernetes release.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>