iku: Add info page for deprecated APIs
There's now an info page for deprecated APIs, that shows
which workloads have been identified to be users of the
deprecated APIs, and what the newest API in the same family is,
if any.

Note that while there shouldn't be any false positives,
there can definitely be false negatives--not all workloads
are created using kubectl apply.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>