iku: Cast bool to str in getters
We use the output from the getters as strings, so cast them
to ensure that we can do string operations such as concatenate,
len, etc.

We could unconditionally cast all unknown types to str,
but it's better to have the program throw and exception to allow
us to detect programming errors easily.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>