ikt: Add "--purge" option to remove-nodes
Since remove-nodes will remove the hosts from the node group
it's impossible to use "ALL" when purging. To avoid this
inconvenience it's nice to have an option that both tears down
and purges the hosts at the same time. This can now be achieved
by using:

ikt remove-hosts --purge <hostnames>

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>