iku: Add Wide/Normal(/Custom) widths
Some list views are quite wide; for people with low:ish resolution
this might be less than ideal. Make it possible to toggle between
predefined widths (Normal and Wide) and allow the user to specify
their own custom view and default to that one if defined, while
still having the other widths as options.

For now only the Pod list view has a alternate widths. The old
one is now defined as the Wide view, and remains the default,
but there's now a view that doesn't show Containers and Pod IP.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>