playbooks/*: Initial Red Hat support
Add initial support for Red Hat (and possibly derivatives).
It has at least one major compromise; to be able to install docker
(which is still the default for older versions of Kubernetes),
it uninstalls runc first.

Also, do note that due to the commercial nature of RHEL, and
the fact that all current users and testers of iKT use Debian
or derivatives, the Red Hat support has seen very little testing.

Bug reports and improvement suggestions (or even patches) are
warmly welcome though!

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>