iku,views: Fixed a lot of corner cases
Fixed a lot of crashes and incorrect behaviour in corner cases
(some introduced by the new view files, some that have been around
for a long time but gone unnoticed until now).

Simplified the log handling code to use more of the common
functions, and removed the support for fields in the
(field_template, override_header) format.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>