iku,views: Fix many corner cases and regressions
Last night's work on the views went on a bit too late,
so a lot of tired mistakes were made.  Hopefully this fixes
most/all of them.

As a bonus values can now be mapped by value or by range.
The first user of this is type field in the events view.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>