iku: Improved search in resource viewer
The resource viewer now shows a bullet in front of all matching
lines, and highlights the current match.

This way there's less ambiguity when there are multiple matches
on one page, and especially if that page is the last page
(since a match previously was just indicated by the matching
line being the first line on screen *except* when it was on
the last page).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>