iku,views: Fix ready/not ready for endpoints
EndpointSlice was using a fancy generator that was intended
to show ready slices with stgroup.OK, and notReady slices
with stgroup.NOT_OK.  Instead it used whatever status the first
item in the list had.

Endpoints didn't even have this kind of highlighting.

Now both have, and both work.  The solution is very ugly though;
it'll be rewritten to use a mapping instead as soon as mappings
can take a secondary value.

Oh, and ellipsise some lists for Service, Endpoints,
and EndpointSlice.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>