iku: Improved resource info and configmap handling
Further improvements for get_resource_info(); we now use
the status field to show when configmaps are empty,
and when references resources are missing (with different
highlighting if they're optional; this isn't ideal from
a accessibility point of view though; it needs revisiting).

If a configmap starts with a script marker (#!), try to
identify the scripting language.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>