iku,curses_helper: Line wrapping for ResourceView
Add line wrapping to the resource viewer.

Config maps are still handled separately; ideally we should
transition to using the resource viewer for that too, but for
now this is an important first step.

The code is getting quite warty; once all string handling has
been unified (currently there are multiple different ways
to represent the same string) we should be able to simplify
all string handling a lot.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>