iku,views: Migrate parts of ConfigMap info view
Part of the ConfigMap info view has now been converted.
We still need some minor additions to be able to reuse
the generic resource viewer instead of the custom viewer
for the ConfigMap data.

For instance the resource viewer should probably have
some way to toggle between formatted/raw data
(or even do the formatting? Maybe it's better to pass
the formatter and the data to the viewer?).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>