iku,parsers,logparser: New/improved parsers
Improve the parsers for openshift-kube-scheduler and prometheus,
and add parsers for kiali, openstack (very, *very* limited),
and multus.

Also fix a corner case in the yaml_line parser (if we have
valid YAML up until the last line of the file we assume, by default,
that the block is valid.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>