views: Massive amount of new+improved view files
Added a lot of new view files, as well as improved a few.
Some of these have an info view too, but mostly it's list view
only.

This should, however, cover all CRDs installed by CRC 2.0,
which is--to say the least--quite a few of them.  There are still
some OpenShift related APIs that are not defined as CRDs, and that
I have no examples for--such APIs still may lack support.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>