iku: Fix view list
The view list was missing all built-in fields,
and behaved in a weird manner if only built-in fields were
used (which is fairly common).

Also add a message that explains why startup is slow.
The real fix is obviously to fix the slow startup, but this
is better than nothing.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>