iku,*: Timeout if the API-server doesn't respond
Default to a 3s timeout on connection if the API-server doesn't
respond. If the API-server is so bogged down that it doesn't
even accept a connection within 3 seconds we don't want to add
to the load unnecessarily.

Besides, on startup or when switching context it's really
confusing if the request takes forever.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>