ikt: Add --forks option
It is now possible to override the number of Ansible forks
from ikt.yaml when preparing, adding, removing, purging,
and upgrading nodes.

The same option has NOT been added to iktadm, since even if we
were to support multiple control planes (not yet supported)
we are unlikely to have more than 5 (the default number of forks).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>