logparser: Remove old-style parsers
Remove all built-in old-style parsers;
this simplifies logparser.py a lot,
and makes things more maintainable (and customisable).

Oh, and removes 700 lines of code...

If this breaks logparsing for something, please either submit
a parser file or an example log with enough extra information
to create a unique match (typically a combination of pod name,
container name, image name).

This also adds a generic parser that only parses timestamps
and glog type entries.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>