iku,schemas,views: View schema, fixed issues found
The schema for views is now more or less complete.
The complexity indicates that the file format for views should
be simplified a bit, but at least it's now possible to detect
the most glaring errors.

Some minor details remain to be fixed (more descriptions,
command line options for the validator to choose whether or not
to exit upon error, etc.)

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>