views: New format for {field,row}_indexes
field_indexes is now of the form:

field_indexes:
  <INDEX_NAME>:
    fields: [list of fields]
    sortcolumn: <SORTCOLUMN>
    sortorder_reverse: <TRUE/FALSE>

row_indexes:
  <INDEX_NAME>:
    fields: [list of fields]

The reason behind this change is that narrow views might
not contain the default sortcolumn used in the default view.
It might also be that the field has a different name in that
view (for instance addresses vs addresses_ellipsised).

This obviously required changes to iku and schemas/vies.json
too. When running validate_yaml a few areas of improvement
were discovered. They were documented in the new tests/TODO.md
file.

TODO was also changed to markdown format (this will eventually
be the case for all TODO/README/FIXME/etc. files).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>