iku,schemas,views: Schema and view fixes
__PodLogViewer has been converted to a view file. This was the last
holdout of the views.

Tightened the schema for views further to only allow a set enum
of generators, availability_checkers, infogetters, listgetters,
objgetters, datagetters, itemgetters, and formatters.

On the iku side of things matching allow lists were added,
and the previous nasty evals were replaced with lookups.
This cuts down on the number of evals significantly--the remaining
ones should be fairly straightforward to replace.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>