iktadm,*: Add upgrade-cni command
Add support for upgrading CNI to iktadm.

This version is (currently) slightly more powerful than
the iku version; this version also upgrades any CNI installers
(and the upgrader is written in a generic enough way that it
could technically upgrade other things than just CNI).

While doing this move download_files() to a separate file to
create a bit more modularity.  This is a function that could
conceivably be useful elsewhere.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>