ikt-install,*: Better fallback handling
If we need to fallback to using PIP to get a python library
we should warn the user and let them decide whether or not this
is acceptable. Also add an option to always disallow this.

While at it improve the commandparser a bit (or at least the
helptexts generated by it).

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>