formatter,*: Refactor and optimise
Move all formatters to a separate file.

Remove all duplicate initialisation of iktconfig; it can be read on
import.  At some point we might want to allow for rereading it, but for
now it's enough that it's read when each program starts.

Compile all regexes used inside loops. Some of this might be
overzealous; most of the loops have few iterations.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>