.pylint: Customisation for our standards
We're explicitly using var == True and var == False in our code,
and don't really intend to change that, so disable that warning.
pylint also seems to think that ikt-install is a bad name. It isn't,
so just disable that warning too.

Finally, for now we're using Python 3.6 as minimum version.
Hence pylint should use that too.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>