iktio.py: Improved file handling + docstrings
Rather than using temporary files as an intermediate step
when we download, just use a BytesIO object until we're sure
whether or not it's a tarfile. Do everything we can in memory
before writing things out to the tempfile.

The tempfile will have secure permissions by default,
we'll change them after writing, and the file move is atomic.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>