Security: Replace shutil.which()
Instead of using shutil.which() directly, we introduce
a wrapper that checks the paths against a list of permitted
paths, makes sure that the path isn't pointing somewhere else
than it should, and that the file and parent directory has safe
owners and permissions.

Also check that the destination path isn't world writable before
running ikt-install.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>