More type hints
Type-hinting for iktprint and curses_helper is on-going,
meaning that a lot of things might complain.

A lot of the strings passed to iktprint have been fixed already,
but many remain, and as for themearray strings only the surface
has been scratched.

The type hints aren't without benefits though, a few corner case
errors have already been uncovered (typically cases of error messages
that would've failed due to having bugs in them...)

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>