Refactor to fix ikt-install issue
Some previous work managed to introduce python3-natsort as
a dependency in a place where it shouldn't be. Due to this
ikt-install failed to work if python3-natsort wasn't already
installed. Since installing that package is one of the purposes
of ikt-install this is obviously not a good thing.

Hopefully things are better off now.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>