iktadm: Check whether options change mid-setup
Since the installation can be done in steps (on purpose),
be interrupted and resumed (on purpose), or fail and be resumed
after fixing the cause for failure, we need to ensure that the
options passed to setup-control-plane are the same as we started
with.

Changing the Pod Network CIDR mid-installation is always
an issue, so abort if this is attempted.
Changing CNI *can* work, so provide an option that permits this,
but default to aborting.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>