curses_helper: A bit more resilience for resize
Handling live-resizing of the terminal is tricky. We try to adjust
in some cases, but notably the windowwidget() doesn't handle
things well at all. Rather than pretend that it does we immediately
error out and tell the user why, instead of exposing them to
an exception.

Also catch exceptions from other troublesome cases. When the content
doesn't fit, don't try to output it.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>