Limit the amount of log lines from dmesg
For containerd and kubelet we already read a limited amount of lines,
but for dmesg we read everything since boot. This turned out to be
a rather bad idea. Starting iku took ages when trying to run it on
a system with 1.3 million lines in dmesg... Now we limit dmesg
to 10k lines and containerd + kubelet to 1000 lines.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>