iku: Add patch_object() and fix some bugs
Now that the patch function in kubernetes_helper has been modified
we can add patch_object() and use it to trigger object patching
from view files.

While doing this some bugs were discovered (if a kind tuple had
the first field == None it'd bypass the None-check and then
trigger an exception), and one case of error handling was
also broken.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>