Start adding early argument and option validation
Rather than having each callback validate the arguments and options
it receives, thus duplicating a lot of code, we can instead do this
validation in the commandparser.

While at it the validation was made stricter (we now check
the proxy URL to be a valid URL, IP-addresses to be valid IP-addresses,
etc.

To avoid missing out on corner cases this introduced a new dependency;
validators.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>