Reformat helptexts a bit
Ideally helptexts should fit within 80 columns. Some of the texts
*almost* fit within this; cmt-install, cmtinv, and cmu now achieves
that goal.

cmt and cmtadm have a long way to go though; the best solution is
probably a partial rewrite of command parser to show commands
and options on a line of their own if the option is longer than
a certain length.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>