views/ClusterPolicy: Fix naming conflict
In case a cluster had both ClusterPolicy.kyverno.io
and ClusterPolicy.nvidia.com installed there would be a naming
conflict and cmu would refuse to start.

The proper solution for this is to index views by their
(Kind, API-family) tuple, but that requires modifying almost 400
files, so for now we just work around it.

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>