cmt: Add --has-data and --format options
Add "--has-data" filter and "--format" output format option
to the api-resources command. "--format" can be used to output
the data in other than the default (table) behaviour.

Notably there's support for outputting entries that can be added
directly to kubernetes_resources.py in the CMT source code,
and space separated lists that can be fed to "kubectl get" to
list all (filtered) resources.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>