cmt: Inventory double-check
In add-node we already double-check so that the node
isn't a part of the cluster, but we don't check whether it's
listed as a node or control plane in the ansible inventory.

Since most playbooks refuse to run on nodes & control planes
(to prevent existing configuration from being overwritten)
a host that shouldn't be marked as a node but is can cause
installation to fail.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>