kubernetes_helper: Add more API-definitions
Add a few (alpha-level) new API-definitions.  They might end up
unused, but maybe having them around might be useful for something?
Who knows. Start by adding a list view for PodSchedulingContext
(this is a rename of the PodScheduling kind, so the view file
is a copy of that one with renames done).

Signed-off-by: David Weinehall <david.weinehall@gmail.com>