networkio: Add a missing flush()
Typically it's OK to rename a file before flushing it
since the move only changes meta-data, but when the move
is across file-system boundaries it silently turns into
a copy, which means that the content is read + written.
To do this the contents of the file needs to have been flushed
first, otherwise the destination file can end up empty.

Signed-off-by: David Weinehall <david.weinehall@intel.com>