cmtadm: Make a note about modifying kubeconfig
When installing rke2 the kubeconfig ends up with 127.0.0.1
as the server IP-address instead of the real IP-address.
This is obviously not ideal if you want worker nodes to be
able to connect or to use the kubeconfig on a remote host.

This will be fixed, but for now just note that the user needs
to fix this themselves.

Signed-off-by: David Weinehall <david.weinehall@intel.com>