Partial bash-completion for cmu
The auto-complete for cmu is missing a lot; it cannot complete
namespace, object, or item, nor can it complete the available
(or even supported) views.

To be able to autocomplete the available/supported views we
need to improve list-views, and to complete namespace/object:item
we need to either implement "cmt get" or use "kubectl get".

Signed-off-by: David Weinehall <david.weinehall@linux.intel.com>