Add bash-completion
Bash-completion. Not perfect (this is my first attempt),
but it's better than nothing.  cmu, in particular,
still has a long way to go, since it currently doesn't
suggest the list of available (or even supported) views;
it just suggests that VIEW is one of the supported arguments.

To enable bash-completion ansithemeprint, cmt, cmu, and cmtinv
were modified to allow disabling of colored output. More
output formats were also added.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>