Add token request/renewal for CRC
Adding token request/renewal makes it no longer necessary
to call oc login before running cmu when using crc clusters.
This isn't generic yet, so it won't work on other clusters
that use tokens, and there are other limitations too,
but it's a start at least.

For now you need to create ~/.kube/credentials (0o600)
with the format:

clusters:
  "clustername":
    contexts:
      "contextname":
         name: "username"
         password: "password"

This format might change though, since the .kube/config file
is made in a way that allows multiple clusters with the same
name; this format does not allow for that.

Once importing crc clusters gets more complete support creating
this file will be automated and better documented; for now
it's only supported on a "it's there if you know" level.

It is currently not possible to pass username + password
on the command line; before supporting this we need a way to strip
the command line; having passwords visible in ps obviously
is a security issue.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>