playbooks: Runtime Endpoint fixes
Add some fixes to handle cases where multiple possible
runtime endpoints are installed on the system.

When, for instance, both containerd and cri-o are installed
on the same system kubeadm won't be able to decide which one
to use, and thus things will fail.

Pass the relevant information explicitly. Also, to help debug
this kind of issues faster next time, output the node
and cluster configuration that will be used during setup.

Signed-off-by: David Weinehall <david.weinehall@intel.com>