cmu,curses_helper: Fix timers
The timers used for asynchronous helpers as well as data refresh
were very imprecise, and on top of that off by a factor 10
(50 seconds instead of 5).

Hopefully this fixes the issue without introducing any new
regressions.

There can still be some variance in case an Ansible job or
REST API request takes a long time, but that's by design.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>