cmu,datagetter: Performance fix
Another performance issue fixed; when fetching a list
with thousands of pods the node status would be fetched
for any pod with status NotReady. This can easily reach
several thousand calls to the API-server if the cluster
is unhealthy.

There are still remaining performance issues that need
to be fixed, but most of the remaining ones will require
threading.

Signed-off-by: David Weinehall <david.weinehall@intel.com>