cmu,*: Improved cluster overview
The cluster overview has seen various improvements.

* The latest 5 events are now included.
* Instead of dividing the space equally between nodes and pods,
  we now give nodes 25% and pods 75% of the space,
  justified by hopefully reasonably expectation that there
  will be several pods per node.
* The total pod and node count is included, as well as
  the selected index value.
* Under the hood the percentage bars are now generated
  as themearrays instead of directly to the screen.

At some point we need to rethink things further. On clusters
scaling to the limits of what Kubernetes can handle, or on smaller
terminal windows, we won't be able to fit all pods, maybe not even
all nodes. Support for scrolling is necessary, and the pod and node
views need to be able to scroll separately.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>