Start adding version cache support
Since upstream Kubernetes now distribute packages using
OBS repositories we can no longer detect new minor versions
of Kubernetes through a repo refresh, nor can we check
what the latest version is that way, since OBS creates
separate repositories for each minor release;
only patch releases are stores in the same repository.

This breaks (at least):

cmtadm prepare
cmtadm upgrade-control-plane
cmt add-node

For this purpose (and maybe other components for future use)
a version-cache directory as well as a playbook to download
files into it has been added.

Whenever the playbook is run it'll check whether the cache has
been updated within the last 6h, and if so skip updates.
If not, or if "force=true" is specified, the release info
for Kubernetes is downloaded.

Currently the information isn't used for anything,
and the only way to use the playbook is to run it manually.

Next step: have cmtadm call the playbook when running "cmtadm cv"
and whenever preparing installation, and cmt whenever running
"cmt add-node", and create repo files based on the information
from these files.  This is high priority since Kubernetes 1.29
will only be available from OBS.

Adding a check for newer upstream versions to the cluster
overview in cmu (and maybe also a more generic version view
that could check things such as versions for various deployments,
binaries, etc.) should also be considered.

Signed-off-by: David Weinehall <david.weinehall@intel.com>