formatter,views: Add Python Traceback formatter
Add a formatter for Python Tracebacks, and use it
to show error messages from RayService.  This is probably
not the correct approach (it feels likely that error
messages from RayService aren't always Tracebacks...),
but that's the only status message reproduced so far,
so let's go with that. Better than nothing at least.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>