requirements.txt: Bump version for validators
Bump the required version for validators; there's
a regex DoS in the URL-checker in any version older than 0.21.
While this is unlikely to affect us (the validators are used to
validate command line options, so any attempts to feed malformed
data would only result in the attacker DoS themselves),
it's good practice to not accept buggy dependencies.

Signed-off-by: David Weinehall <david.weinehall@intel.com>