cmu,*: Add a theme for colour vision deficiency
The cluster overview relies heavily on colour coding
to display node and pod status as well as the different
types of memory & cpu usage.

Add a theme that starts addressing this. There are still
things that need to be remedied elsewhere in cmu, mainly
the log viewer that currently relies on colour coding
for severity, but that needs to be handled slightly differently.

If anyone with colour vision deficiency has suggestions
how to improve things further, consider filing a feature request.

Signed-off-by: David Weinehall <david.weinehall@intel.com>