cmt: Confirm remove-nodes
It feels obvious in hindsight that removing nodes should
require confirmation...

Add code to confirm removal of nodes, as well as a summary
of what nodes will be removed "hidden" behind --verbose
(the user typically just passed in the list of nodes,
so they don't necessarily want to see it by default).

To facilitate cases where a lot of nodes are deleted at
once we add list chunking. This should eventually be used
in more places where we list nodes, such as preparing hosts,
adding nodes, upgrading nodes, and purging hosts.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>