cmu,formatter: Fixes for markdown formatter
The Markdown formatter is, to say the least, not very good.
We should definitely abandon the current half-measure
and use some pre-existing implementation instead.

But until then at least patch things up a bit.

For now there are no users of the Markdown viewer,
but that will hopefully change soon:ish.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>