cmu,*: (Hopefully) fix asynchronous behaviour
An SSLContext cannot be shared between contexts in a threadsafe
manner, so we have to create separate contexts for each thread.
Do this in a nice:ish (but still ugly) way by creating
a PoolManagerContext() class and using:

With PoolManagerContext(...) around any code that needs
a PoolManager (or ProxyManager).

Testing seems to indicate that this is enough to fix the crashes
observed when fetching data asynchronously.  There are still
things we can improve; for instance we should probably
implement backoff in case requests are starting to take too long
or if they fail.

To (try to) minimise the risk of regressions we also add
a unit-test for asynchronous fetches.

This commit also fixes listgetter_namespaced_resources()
to fetch resources in parallel. We're attempting to fetch all
objects in a namespace from all resources that CMT knows of,
which is several hundred. This can thus potentially take a *really*
long time (ideally we should add a notice about this, but that
will need a bit of UI improvement since there's currently no way
to add warnings about potentially slow listgetters).

To avoid overloading the API-server we don't use as many workers
as there are namespaced resources though.

There are still FIXMEs remaining before we can merge this
to the main branch; proper comments and type-annotations,
for instance. We also need to convert the context and inventory
view to use the same type of asynchronous behaviour.

The new ReExecutor() class should also be moved outside cmu,
since it can (potentially) be useful elsewhere, and having it
separately would enable it to be used from unit-tests.

A more pressing issue is properly shutting down the executor
when switching views though.  We currently shut it down
when switching via the menu, but when following objects
(say, from Cluster Overview to a Pod) we don't.  This needs
to be solved before enabling asynchronous data fetches
for the rest of the UI.

This must also be fixed before merging to the main branch.
Having a global executor would solve this (we can then flush
items from the worker unless they're tagged persistent or something
such; this would also allow for tasks that run all the time,
should we need any such tasks).

Signed-off-by: David Weinehall <david.weinehall@gmail.com>