cmu,*: Convert Context & Inventory to new async
These conversions are not exactly clean, but it lets
us get rid of most of the special cases for these two views.

To properly allow for the necessary async behaviour we probably
need to either create some better wrappers, or to make the
arguments passed to listgetter_async more comprehensive.

There are still changes that need to be made; if the list of hosts
changes (hosts added or removed from the inventory) we currently
do not update the list of hosts we ping, so their status remains
unknown.

The same goes for contexts. This is probably less of an issue though.
In the case of contexts it's probably enough for now to just
document it as a known limitation.

ReExecutor() is now split into a file of its own. Adapt
the unit-test accordingly.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>