cmu: Fix asynchronous Inventory behaviour
We now detect addition and deletion of hosts.
Rather than just remove them from the list they'll
be marked as deleted and dimmed.

The next step is to do the same with all asynchronous list
items, and trigger removal only when forcing update [F5],
and to make deleted items unselectable (take care to handle
the special case where none of the entries are selectable).

Signed-off-by: David Weinehall <david.weinehall@intel.com>