cmttypes.py: Augment custom Exceptions
Augment our custom exceptions with extra information,
for use in error messages, logs, etc.

TODO: Currently there's duplication; we should have a base
class to inherit from instead. We should also wrap the rest
of the exceptions we throw.

This way we'll be able to throw exceptions in libraries
instead of using sys.exit() and still get nicely
formatted error messages when needed.

Signed-off-by: David Weinehall <david.weinehall@intel.com>