curses_helper.py: Start using custom errors
Start switching to using ProgrammingError whenever we know
it's such, and provide the data that we previously (erroneously)
tried to log using cmtlog.py.  This way we can, if we want,
propagate the exception all the way upp to the calling
executable and log things there (where the logging belongs).

Signed-off-by: David Weinehall <david.weinehall@gmail.com>