cmtlib.py,cmu, generator.py: Refactoring
Move some code from cmu to generator.py;
ideally none of the code that formats input => output
should be in cmu. The eventual goal for both cmu and cmt
to be able to use the same formatters (except the output
bit that should output either ANSIThemeString or ThemeString).

cmu should pretty much just take YAML-files that describe views,
log viewer rules.

To that point we've got a loooong way to go though.
But if we refactor a bit of code here, a fragment of code there,
we hopefully slowly get closer to that goal.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>