views/__VersionData,*: Add version view
Add a view that (currently) shows that latest version
of Kubernetes and the installed version. If opening
the info view the changelog will be shown.

Later on the intention is to add support for more
upstream components than just Kubernetes.

Also it seems that the release page for Kubernetes
is broken; it still doesn't (as of doing this commit)
list Kubernetes 1.29.2, which is currently the latest
version.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>