*: Use KubernetesResourceCache with use-testdata
We're reading the testdata from disk; this can potentially
be extremely slow. Cache the requests, so that we only need
to do such reads once per resource.

There are still some really slow corner cases left,
(the CNI view and the Namespace info view, for instance),
but on the whole things are definitely better now than before.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>