*: Implement version check in code
We had an Ansible playbook that would fetch version
data for Kubernetes, but due to a bug in Ubuntu 22.04 LTS
we cannot use native Jinja2, and thus we cannot parse
the content properly.  Rather than try to come up with
some ugly workaround we bite the bullet and fetch
things with download_files() instead.

Hopefuly this fixes installation issues on Ubuntu 22.04 LTS
without introducing regressions elsewhere.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>