*: Remove dependency on ansithemeprint from cmtio
cmtio.py is used by ansithemeprint.py to load themefiles.
Hence it's not really a good idea to use ansithemeprint()
from cmtio (well, ideally we shouldn't print *anything*
from cmtio, since it's a library, but that's a concern
for another day).  Just use regular print() instead.

It's not quite as nice, but it'll have to do.  Eventually
we might shift to raising exceptions for errors, and to use
logging for debug messages.

Signed-off-by: David Weinehall <david.weinehall@intel.com>